+++ experimental/packages/opentelemetry-exporter-metrics-otlp-proto/karma.conf.js
@@ -14,4 +14,15 @@
  * limitations under the License.
  */
 
-export { OTLPMetricExporter } from './platform';
+const karmaWebpackConfig = require('../../../karma.webpack');
+const karmaBaseConfig = require('../../../karma.base');
+
+module.exports = config => {
+  config.set(
+    Object.assign({}, karmaBaseConfig, {
+      webpack: karmaWebpackConfig,
+      files: ['test/browser/index-webpack.ts'],
+      preprocessors: { 'test/browser/index-webpack.ts': ['webpack'] },
+    })
+  );
+};