+++ experimental/CHANGELOG.md
@@ -11,6 +11,8 @@ For notes on migrating to 2.x / 0.200.x see [the upgrade guide](doc/upgrade-to-2
 ### :rocket: Features
 * feat(exporter-otlp-*): update proto to `v1.7.0`
 
+* feat(exporter-metrics-otlp-proto): Support to protobuf in browser metrics. [#5710](https://github.com/open-telemetry/opentelemetry-js/pull/5710) @YangJonghun
+
 ### :bug: Bug Fixes
 
 ### :books: Documentation